Correção Exe03

+++ src/aula01/exercicio3/Exe03.java
@@ -8,11 +8,11 @@ public class Exe03 {
 
         System.out.println("---------Maior, Menor e Média de 3 Números---------");
         System.out.print("Digite o primeiro número: ");
-        primeiroNumero = ler.nextInt();
+        primeiroNumero = ler.nextDouble();
         System.out.print("Digite o segundo número: ");
-        segundoNumero = ler.nextInt();
+        segundoNumero = ler.nextDouble();
         System.out.print("Digite o terceiro número: ");
-        terceiroNumero = ler.nextInt();
+        terceiroNumero = ler.nextDouble();
 
         double maior = Math.max(primeiroNumero, Math.max(segundoNumero, terceiroNumero));
         double menor = Math.min(primeiroNumero, Math.min(segundoNumero, terceiroNumero));
